Commit summary:
- Move evaluate without marabou from MarabouNetwork.py to MarabouNetworkNNet.py

- Consolidate all evaluate nnet methods into one

- Remove evaluate multiple

- Remove tests without asserts

- Change comparisons of test results from rounding to comparison up to a tolerance factor
